WIP: More refactoring
