+++ mne/defaults.py
@@ -31,6 +31,10 @@ DEFAULTS = dict(
                 misc='misc', seeg='sEEG', dipole='Dipole', eog='EOG',
                 gof='Goodness of fit', ecg='ECG', emg='EMG', bio='BIO',
                 ecog='ECoG', hbo='Oxyhemoglobin', hbr='Deoxyhemoglobin'),
+    titles_short=dict(eeg='EEG', grad='Grad', mag='Mag',
+                      misc='Misc', seeg='sEEG', dipole='Dip', eog='EOG',
+                      gof='GOF', ecg='ECG', emg='EMG', bio='BIO',
+                      ecog='ECoG', hbo='Oxy', hbr='Deoxy'),
     mask_params=dict(marker='o',
                      markerfacecolor='w',
                      markeredgecolor='k',
